railway-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.3.1
python-dotenv==1.0.1
supabase==2.4.6
anthropic==0.29.0
upstash-redis==1.1.0
